Show worldcoin login flow on index
