+++ packages/nextjs/package.json
@@ -19,6 +19,7 @@
     "@heroicons/react": "^2.0.11",
     "@rainbow-me/rainbowkit": "^0.11.0",
     "@uniswap/sdk": "^3.0.3",
+    "@worldcoin/idkit": "^0.5.1",
     "daisyui": "^2.31.0",
     "ethers": "^5.0.0",
     "next": "^13.1.6",
